pyproject.toml
[project]
name = "sweetiebot"
version = "0.1.1"
description = "Sweetie Bot framework extensions"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "pydantic>=2.6",
  "pyyaml>=6.0",
]

[project.scripts]
sweetiebot = "sweetiebot.cli.main:main"

[project.entry-points."sweetiebot.plugins"]
local_dialogue_provider = "sweetiebot.plugins.builtins:LocalDialogueProviderPlugin"
demo_routine_pack = "sweetiebot.plugins.builtins:DemoRoutinePackPlugin"
default_safety_policy = "sweetiebot.plugins.builtins:DefaultSafetyPolicyPlugin"
